+++ Stripe/PublicHeaders/STPRedirectContext.h
@@ -37,22 +37,44 @@ typedef NS_ENUM(NSUInteger, STPRedirectContextState) {
 };
 
 /**
- A callback run when the context believes the redirect action has been completed.
+ A callback that is executed when the context believes the redirect action has been completed.
 
  @param sourceID The stripe id of the source.
  @param clientSecret The client secret of the source.
  @param error An error if one occured. Note that a lack of an error does not 
  mean that the action was completed successfully, the presence of one confirms 
  that it was not. Currently the only possible error the context can know about 
- is if SFSafariViewController fails its initial load (like the user has no 
+ is if SFSafariViewController fails its initial load (e.g. the user has no
  internet connection, or servers are down).
  */
-typedef void (^STPRedirectContextCompletionBlock)(NSString *sourceID, NSString * __nullable clientSecret, NSError * __nullable error);
+typedef void (^STPRedirectContextSourceCompletionBlock)(NSString *sourceID, NSString * __nullable clientSecret, NSError * __nullable error);
 
 /**
- This is a helper class for handling redirect sources.
+ A callback that is executed when the context believes the redirect action has been completed.
 
- Init and retain an instance with the redirect flow source you want to handle,
+ This type has been renamed to `STPRedirectContextSourceCompletionBlock` and deprecated.
+ */
+__attribute__((deprecated("STPRedirectContextCompletionBlock has been renamed to STPRedirectContextSourceCompletionBlock", "STPRedirectContextSourceCompletionBlock")))
+typedef STPRedirectContextSourceCompletionBlock STPRedirectContextCompletionBlock;
+
+
+/**
+ A callback that is executed when the context believes the redirect action has been completed.
+
+ @param clientSecret The client secret of the PaymentIntent
+ @param error An error if one occured. Note that a lack of an error does not
+ mean that the action was completed successfully, the presence of one confirms
+ that it was not. Currently the only possible error the context can know about
+ is if SFSafariViewController fails its initial load (e.g. the user has no
+ internet connection, or servers are down).
+ */
+typedef void(^STPRedirectContextPaymentIntentCompletionBlock)(NSString *clientSecret, NSError * __nullable error);
+
+/**
+ This is a helper class for handling redirects associated with STPSource and
+ STPPaymentIntents.
+
+ Init and retain an instance with the Source or PaymentIntent you want to handle,
  then choose a redirect method. The context will fire the completion handler
  when the redirect completes.
 
@@ -65,13 +87,13 @@ typedef void (^STPRedirectContextCompletionBlock)(NSString *sourceID, NSString *
  have not actually completed the necessary actions to authorize the charge.
 
  You should not use either this class, nor `STPAPIClient`, as a way
- to determine when you should charge the source or to determine if the redirect
- was successful. Use Stripe webhooks on your backend server to listen for source
+ to determine when you should charge the Source or to determine if the redirect
+ was successful. Use Stripe webhooks on your backend server to listen for Source
  state changes and to make the charge.
  
  See https://stripe.com/docs/sources/best-practices
  */
-NS_EXTENSION_UNAVAILABLE("Redirect based sources are not available in extensions")
+NS_EXTENSION_UNAVAILABLE("STPRedirectContext is not available in extensions")
 @interface STPRedirectContext : NSObject
 
 /**
@@ -80,7 +102,7 @@ NS_EXTENSION_UNAVAILABLE("Redirect based sources are not available in extensions
 @property (nonatomic, readonly) STPRedirectContextState state;
 
 /**
- Initializer for context.
+ Initializer for context from an `STPSource`.
 
  @note You must ensure that the returnURL set up in the created source
  correctly goes to your app so that users can be returned once
@@ -90,15 +112,34 @@ NS_EXTENSION_UNAVAILABLE("Redirect based sources are not available in extensions
  @param completion A block to fire when the action is believed to have 
  been completed.
 
- @return Nil if the specified source is not a redirect-flow source. Otherwise 
+ @return nil if the specified source is not a redirect-flow source. Otherwise
  a new context object.
 
- @note Firing of the completion block does not necessarily mean the user 
+ @note Execution of the completion block does not necessarily mean the user
  successfully performed the redirect action. You should listen for source status
  change webhooks on your backend to determine the result of a redirect.
  */
 - (nullable instancetype)initWithSource:(STPSource *)source
-                             completion:(STPRedirectContextCompletionBlock)completion;
+                             completion:(STPRedirectContextSourceCompletionBlock)completion;
+
+/**
+ Initializer for context from an `STPPaymentIntent`.
+
+ This should be used when the `status` is `STPPaymentIntentStatusRequiresSourceAction`.
+ If the next action involves a redirect, this init method will return a non-nil object.
+
+ @param paymentIntent The STPPaymentIntent that needs a redirect.
+ @param completion A block to fire when the action is believed to have
+ been completed.
+
+ @return nil if the provided PaymentIntent does not need a redirect. Otherwise
+ a new context object.
+
+ @note Execution of the completion block does not necessarily mean the user
+ successfully performed the redirect action.
+ */
+- (nullable instancetype)initWithPaymentIntent:(STPPaymentIntent *)paymentIntent
+                                    completion:(STPRedirectContextPaymentIntentCompletionBlock)completion;
 
 /**
  Use `initWithSource:completion:`
@@ -109,7 +150,7 @@ NS_EXTENSION_UNAVAILABLE("Redirect based sources are not available in extensions
  Starts a redirect flow.
 
  You must ensure that your app delegate listens for  the `returnURL` that you
- set on your source object, and forwards it to the Stripe SDK so that the
+ set on the Stripe object, and forwards it to the Stripe SDK so that the
  context can be notified when the redirect is completed and dismiss the
  view controller. See `[Stripe handleStripeURLCallbackWithURL:]`
 
@@ -122,7 +163,7 @@ NS_EXTENSION_UNAVAILABLE("Redirect based sources are not available in extensions
  over the redirect method, you can use `startSafariViewControllerRedirectFlowFromViewController` 
  or `startSafariAppRedirectFlow`
  
- If the source supports a native app, and that app is is installed on the user's
+ If the redirect supports a native app, and that app is is installed on the user's
  device, this call will do a direct app-to-app redirect instead of showing
  a web url. 
 
@@ -139,7 +180,7 @@ NS_EXTENSION_UNAVAILABLE("Redirect based sources are not available in extensions
  from the passed in view controller.
 
  You must ensure that your app delegate listens for  the `returnURL` that you
- set on your source object, and forwards it to the Stripe SDK so that the
+ set on the Stripe object, and forwards it to the Stripe SDK so that the
  context can be notified when the redirect is completed and dismiss the
  view controller. See `[Stripe handleStripeURLCallbackWithURL:]`