+++ Stripe/STPRedirectContext.m
@@ -7,9 +7,11 @@
 //
 
 #import "STPRedirectContext.h"
+#import "STPRedirectContext+Private.h"
 
 #import "STPBlocks.h"
 #import "STPDispatchFunctions.h"
+#import "STPPaymentIntent.h"
 #import "STPSource.h"
 #import "STPURLCallbackHandler.h"
 #import "STPWeakStrongMacros.h"
@@ -22,14 +24,6 @@ NS_ASSUME_NONNULL_BEGIN
 typedef void (^STPBoolCompletionBlock)(BOOL success);
 
 @interface STPRedirectContext () <SFSafariViewControllerDelegate, STPURLCallbackListener>
-/// Optional URL for a native app. This is passed directly to `UIApplication openURL:`, and if it fails this class falls back to `redirectUrl`
-@property (nonatomic, nullable, copy) NSURL *nativeRedirectUrl;
-/// The URL to redirect to, assuming `nativeRedirectUrl` is nil or fails to open. Cannot be nil if `nativeRedirectUrl` is.
-@property (nonatomic, nullable, copy) NSURL *redirectUrl;
-/// The expected `returnUrl`, passed to STPURLCallbackHandler
-@property (nonatomic, copy) NSURL *returnUrl;
-/// Completion block to execute when finished redirecting, with optional error parameter.
-@property (nonatomic, copy) STPErrorBlock completion;
 
 @property (nonatomic, strong, nullable) SFSafariViewController *safariVC;
 @property (nonatomic, assign, readwrite) STPRedirectContextState state;
@@ -43,7 +37,7 @@ typedef void (^STPBoolCompletionBlock)(BOOL success);
 @implementation STPRedirectContext
 
 - (nullable instancetype)initWithSource:(STPSource *)source
-                             completion:(STPRedirectContextCompletionBlock)completion {
+                             completion:(STPRedirectContextSourceCompletionBlock)completion {
 
     if (source.flow != STPSourceFlowRedirect
         || !(source.status == STPSourceStatusPending ||
@@ -58,6 +52,30 @@ typedef void (^STPBoolCompletionBlock)(BOOL success);
                                     completion(source.stripeID, source.clientSecret, error);
                                 }];
     return self;
+}
+
+- (nullable instancetype)initWithPaymentIntent:(STPPaymentIntent *)paymentIntent
+                                    completion:(STPRedirectContextPaymentIntentCompletionBlock)completion {
+    if (!(paymentIntent.returnUrl != nil
+          && paymentIntent.status == STPPaymentIntentStatusRequiresSourceAction
+          && [paymentIntent.allResponseFields[@"next_source_action"] isKindOfClass: [NSDictionary class]])) {
+        return nil;
+    }
+
+    NSDictionary *nextSourceAction = paymentIntent.allResponseFields[@"next_source_action"];
+    if (!([nextSourceAction[@"type"] isEqual:@"authorize_with_url"]
+          && [nextSourceAction[@"value"] isKindOfClass:[NSDictionary class]]
+          && [nextSourceAction[@"value"][@"url"] isKindOfClass:[NSString class]])) {
+        return nil;
+    }
+
+    NSString *redirectUrl = nextSourceAction[@"value"][@"url"];
+    return [self initWithNativeRedirectUrl:nil
+                               redirectUrl:[NSURL URLWithString:redirectUrl]
+                                 returnUrl:paymentIntent.returnUrl
+                                completion:^(NSError * _Nullable error) {
+                                    completion(paymentIntent.clientSecret, error);
+                                }];
 }
 
 /**