+++ Tests/Tests/STPFixtures.h
@@ -15,6 +15,8 @@ extern NSString *const STPTestJSONCustomer;
 
 extern NSString *const STPTestJSONCard;
 
+extern NSString *const STPTestJSONPaymentIntent;
+
 extern NSString *const STPTestJSONSource3DS;
 extern NSString *const STPTestJSONSourceAlipay;
 extern NSString *const STPTestJSONSourceBancontact;
@@ -126,6 +128,11 @@ extern NSString *const STPTestJSONSourceSOFORT;
  A Source object with type Alipay and a native redirect url
  */
 + (STPSource *)alipaySourceWithNativeUrl;
+
+/**
+ A PaymentIntent object
+ */
++ (STPPaymentIntent *)paymentIntent;
 
 /**
  A PaymentConfiguration object with a fake publishable key. Use this to avoid